🚚 [rendering] Rename package `cutty.{templates => rendering}`
