+++ src/cutty/rendering/adapters/cookiecutter/__init__.py
