+++ src/cutty/rendering/adapters/cookiecutter/extensions.py
