+++ src/cutty/rendering/domain/renderbind.py
@@ -1,7 +1,7 @@
 """Binding variables."""
 from collections.abc import Sequence
 
-from cutty.templates.domain.render import Renderer
+from cutty.rendering.domain.render import Renderer
 from cutty.variables.domain.binders import Binder
 from cutty.variables.domain.bindings import Binding
 from cutty.variables.domain.variables import Variable
