+++ src/cutty/rendering/domain/renderfiles.py
@@ -6,7 +6,7 @@ from collections.abc import Sequence
 from cutty.filestorage.domain.files import File
 from cutty.filestorage.domain.files import loadfile
 from cutty.filesystems.domain.path import Path
-from cutty.templates.domain.render import Renderer
+from cutty.rendering.domain.render import Renderer
 from cutty.variables.domain.bindings import Binding
