+++ tests/unit/conftest.py
@@ -1,9 +1,9 @@
 """Unit test fixtures for cutty."""
 import pytest
 
-from cutty.templates.domain.render import createrenderer
-from cutty.templates.domain.render import defaultrenderregistry
-from cutty.templates.domain.render import Renderer
+from cutty.rendering.domain.render import createrenderer
+from cutty.rendering.domain.render import defaultrenderregistry
+from cutty.rendering.domain.render import Renderer
 from cutty.variables.domain.variables import GenericVariable
