+++ tests/unit/rendering/adapters/cookiecutter/__init__.py
@@ -1,0 +1,1 @@
+"""Unit tests for cutty.rendering.adapters.cookiecutter."""
