+++ tests/unit/rendering/adapters/cookiecutter/test_render.py
@@ -1,4 +1,4 @@
-"""Unit tests for cutty.templates.adapters.cookiecutter.render."""
+"""Unit tests for cutty.rendering.adapters.cookiecutter.render."""
 from collections.abc import Callable
 from typing import Any
 
@@ -8,9 +8,9 @@ from cutty.filestorage.domain.files import RegularFile
 from cutty.filesystems.adapters.dict import DictFilesystem
 from cutty.filesystems.domain.path import Path
 from cutty.filesystems.domain.purepath import PurePath
-from cutty.templates.adapters.cookiecutter.render import CookiecutterConfig
-from cutty.templates.adapters.cookiecutter.render import createcookiecutterrenderer
-from cutty.templates.domain.render import Renderer
+from cutty.rendering.adapters.cookiecutter.render import CookiecutterConfig
+from cutty.rendering.adapters.cookiecutter.render import createcookiecutterrenderer
+from cutty.rendering.domain.render import Renderer
 from cutty.variables.domain.bindings import Binding
