+++ tests/unit/rendering/adapters/test_prompts.py
