+++ tests/unit/rendering/domain/test_bindings.py
