+++ tests/unit/rendering/domain/test_variables.py
