--- tests/unit/templates/__init__.py
@@ -1,1 +1,0 @@
-"""Unit tests for cutty.templates."""
