--- tests/unit/templates/adapters/__init__.py
@@ -1,1 +1,0 @@
-"""Unit tests for cutty.templates.adapters."""
