--- tests/unit/templates/adapters/cookiecutter/__init__.py
@@ -1,1 +1,0 @@
-"""Unit tests for cutty.templates.adapters.cookiecutter."""
